complete category crud

+++ routes/category.js
@@ -1,25 +1,29 @@
 // @ts-check
-const { collections, fetchNextId } = require("../models/mongo");
+const {
+  collections,
+  fetchNextId,
+  simpleInsertOne,
+  workInTransaction,
+} = require("../models/mongo");
 const { CategorySchema } = require("../models/category.model");
+const checkParamsIdExists = require("../middleware/check-params-id-exists");
 const validatePipe = require("../middleware/validate-pipe");
+const loginCheck = require("../middleware/login-check");
 const router = require("express").Router();
 
 // 假設已經 connectDB
 const category_coll = collections.category;
 
 // GET from database
-router.get(
-  "/",
-  function (req, res) {
-    category_coll
-      .find(req.query)
-      // .sort({ categoryId: 1 })
-      .toArray(function (err, result) {
-        if (err) throw err;
-        res.status(200).send(result);
-      });
-  }
-);
+router.get("/", function (req, res) {
+  category_coll
+    .find(req.query)
+    // .sort({ categoryId: 1 })
+    .toArray(function (err, result) {
+      if (err) throw err;
+      res.status(200).send(result);
+    });
+});
 
 // GET certain data from database
 router.get("/:id", function (req, res) {
@@ -31,40 +35,43 @@ router.get("/:id", function (req, res) {
 });
 
 // Post the info
-router.post("/", validatePipe("body", CategorySchema), async function (
-  req,
-  res
-) {
-  const postData = {
-    _id: await fetchNextId(category_coll.collectionName),
-    ...req.body,
-  };
-  category_coll.insertOne(postData, function (err, result) {
-    if (err) throw err;
-    console.log("1 category info inserted.");
-    res.status(201).send(result.ops[0]);
-  });
+router.post(
+  "/",
+  validatePipe("body", CategorySchema),
+  loginCheck(category_coll),
+  async function (req, res) {
+    /*
+    const postData = {
+      _id: await fetchNextId(category_coll.collectionName),
+      ...req.body,
+    };
 
-  /*
-  category_coll.countDocuments(function (err, count) {
-    category_coll
-      .find({})
-      .sort({ categoryId: 1 })
-      .toArray(function (err, result) {
-        let id = 0;
-        if (count != 0) {
-          id = result[count - 1].categoryId;
-        }
-        const postData = {
-          categoryId: id + 1,
-          categoryName: req.body.categoryName,
-        };
-      });
-  });
-  */
-});
+    category_coll.insertOne(postData, function (err, result) {
+      if (err) throw err;
+      console.log("1 category info inserted.");
+      res.status(201).send(result.ops[0]);
+    });
+    */
+    workInTransaction(async (session) => {
+      const result = await simpleInsertOne(
+        collections.category,
+        { name: req.body.name, userId: req.userId },
+        session
+      );
+
+      if (req.body.hashtags) {
+        await collections.user.updateOne(
+          { _id: req.userId },
+          { $set: { [`categoryTags.${result.insertedId}`]: req.body.hashtags } }
+        );
+      }
+      res.status(201).json(result.ops[0]);
+    });
+  }
+);
 
 // PUT to update certain row info
+/*
 router.put("/:id", validatePipe("body", CategorySchema), function (req, res) {
   const putFilter = {
     _id: req.params.id,
@@ -84,12 +91,40 @@ router.put("/:id", validatePipe("body", CategorySchema), function (req, res) {
     }
   );
 });
+*/
 
 // PATCH to update certain row info
 router.patch(
   "/:id",
   validatePipe("body", CategorySchema, { context: { partial: true } }),
+  loginCheck(category_coll),
+  checkParamsIdExists(category_coll),
   function (req, res) {
+    const category = req.convert_from_params.id;
+    if (category.userId !== req.userId)
+      return res.status(403).json("No access");
+
+    workInTransaction(async (session) => {
+      if (req.body.name) {
+        await collections.category.updateOne(
+          { _id: req.params.id },
+          { $set: { name: req.body.name } },
+          { session }
+        );
+      }
+
+      if (req.body.hashtags) {
+        await collections.user.updateOne(
+          { _id: req.userId },
+          { $set: { [`categoryTags.${req.params.id}`]: req.body.hashtags } },
+          { session }
+        );
+      }
+
+      res.status(200).json("update success");
+    });
+
+    /*
     const patchFilter = { _id: req.params.id };
     const patchData = {
       $set: req.body,
@@ -105,22 +140,37 @@ router.patch(
         res.status(200).send(result.value);
       }
     );
+    */
   }
 );
 
 // DELETE certain row
-router.delete("/:id", function (req, res) {
-  const deleteFilter = {
-    _id: req.params.id,
-  };
-  category_coll.deleteOne(deleteFilter, (err, result) => {
-    console.log(
-      `Delete row: ${req.params.id} with filter: ${deleteFilter}. Deleted: ${result.result.n}`
-    );
-    res
-      .status(200)
-      .send("Delete row: " + req.params.id + " from db Successfully!");
-  });
-});
+router.delete(
+  "/:id",
+  loginCheck(category_coll),
+  checkParamsIdExists(category_coll),
+  async function (req, res) {
+    const category = req.convert_from_params.id;
+    if (category.userId !== req.userId)
+      return res.status(403).json("No access");
+
+    await collections.category.deleteOne({ _id: req.params.id });
+    res.status(200).json("Delete row: " + req.params.id + " from db Successfully!");
+
+    /*
+    const deleteFilter = {
+      _id: req.params.id,
+    };
+    category_coll.deleteOne(deleteFilter, (err, result) => {
+      console.log(
+        `Delete row: ${req.params.id} with filter: ${deleteFilter}. Deleted: ${result.result.n}`
+      );
+      res
+        .status(200)
+        .send("Delete row: " + req.params.id + " from db Successfully!");
+    });
+    */
+  }
+);
 
 module.exports = router;
